email_relay/email_to_pubsub/requirements.txt
google-cloud-pubsub
aiosmtpd
